Fixed a layout bug.
Essentially when a user specifies a popover's content size they expect that to be the size
of the content view not the entire pop-over. However, the size of the content view was
unpredictable because it depended on the position of the arrow. That's fixed now.

Note that the actual size still differs by the size of the border but it's predictable now