allow public readonly access to FPTouch & FPPopoverView
perhaps text-fields weren't anticipated in a small popover, but in the
case where a text-field occurs, if the popover is small enough, it can
still fit entirely in the window above the keyboard … if it can be
manipulated.  by adding access to these items, it can be easily
manipulated with 3rd party DAKeyboardControl category.